# docs/api-docs.json
{"openapi":"3.0.1","info":{"title":"","description":""},"externalDocs":{"description":"","url":""},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"記事API"}],"paths":{"/api/category/{id}":{"put":{"tags":["category-controller"],"operationId":"editCategory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.editcategory.request.EditCategoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.editcategory.response.EditCategoryResponse"}}}}}},"delete":{"tags":["category-controller"],"operationId":"deleteCategory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.deletecategory.response.DeleteCategoryResponse"}}}}}}},"/api/article/{id}":{"get":{"tags":["記事API"],"summary":"単一記事取得API","description":"指定した記事IDに紐づく記事情報の取得と返却を行う","operationId":"getArticleById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","description":"記事ID"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.getarticlebyid.response.GetArticleByIdResponse"}}}}}},"put":{"tags":["記事API"],"summary":"記事編集API","description":"登録済み記事情報の変更を行う","operationId":"editArticle","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","description":"記事ID"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.editarticle.request.EditArticleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.editarticle.response.EditArticleResponse"}}}}}},"delete":{"tags":["記事API"],"summary":"記事削除API","description":"登録済み記事情報の削除を行う","operationId":"deleteArticle","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","description":"記事ID"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.deletearticle.response.DeleteArticleResponse"}}}}}}},"/api/user/signup":{"post":{"tags":["user-controller"],"operationId":"signup","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.user.signup.request.UserSignupRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/file":{"post":{"tags":["file-controller"],"operationId":"postFile","parameters":[{"name":"requestedDateTime","in":"query","required":true,"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.domain.datetime.RequestedDateTime"}}],"requestBody":{"content":{"application/json":{"schema":{"required":["multipartFile"],"type":"object","properties":{"multipartFile":{"type":"string","format":"binary"}}}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.file.postfile.response.PostFileResponse"}}}}}}},"/api/category":{"get":{"tags":["category-controller"],"operationId":"getCategory","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.getcategory.response.GetCategoryResponse"}}}}}},"post":{"tags":["category-controller"],"operationId":"registerCategory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.registercategory.request.RegisterCategoryRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.registercategory.response.RegisterCategoryResponse"}}}}}}},"/api/article":{"get":{"tags":["記事API"],"summary":"記事取得API","description":"公開済み記事情報の取得と返却を行う","operationId":"getArticle","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.getarticle.response.GetArticleResponse"}}}}}},"post":{"tags":["記事API"],"summary":"記事登録API","description":"新規記事の登録を行う","operationId":"registerArticle","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.registerarticle.request.RegisterArticleRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.registerarticle.response.RegisterArticleResponse"}}}}}}},"/file/{fileName}":{"get":{"tags":["file-controller"],"operationId":"getFile","parameters":[{"name":"fileName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"type":"string","format":"byte"}}}}}}}},"/api/user/auth":{"get":{"tags":["user-controller"],"operationId":"auth","responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"jp.gr.java_conf.simpleblogapi.presentation.category.editcategory.request.EditCategoryRequest":{"type":"object","properties":{"name":{"type":"string"},"parentId":{"type":"integer","format":"int32"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.editcategory.response.CategoryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"parentId":{"type":"integer","format":"int32"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.editcategory.response.EditCategoryResponse":{"type":"object","properties":{"result":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.editcategory.response.ResultResponse"},"error":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.editcategory.response.ErrorResponse"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.editcategory.response.ErrorResponse":{"type":"object","properties":{"message":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.editcategory.response.ResultResponse":{"type":"object","properties":{"category":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.editcategory.response.CategoryResponse"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.editarticle.request.EditArticleRequest":{"type":"object","properties":{"title":{"type":"string"},"categoryId":{"type":"integer","format":"int32"},"description":{"type":"string"},"text":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.editarticle.response.ArticleResponse":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"categoryId":{"type":"integer","format":"int32"},"categoryName":{"type":"string"},"createDateTime":{"type":"string","format":"date-time"},"updateDateTime":{"type":"string","format":"date-time"},"description":{"type":"string"},"text":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.editarticle.response.EditArticleResponse":{"type":"object","properties":{"result":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.editarticle.response.ResultResponse"},"error":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.editarticle.response.ErrorResponse"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.editarticle.response.ErrorResponse":{"type":"object","properties":{"message":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.editarticle.response.ResultResponse":{"type":"object","properties":{"article":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.editarticle.response.ArticleResponse"}}},"jp.gr.java_conf.simpleblogapi.presentation.user.signup.request.UserSignupRequest":{"type":"object","properties":{"userId":{"type":"string"},"password":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.domain.datetime.RequestedDateTime":{"type":"object","properties":{"value":{"type":"string","format":"date-time"}}},"jp.gr.java_conf.simpleblogapi.presentation.file.postfile.response.PostFileResponse":{"type":"object","properties":{"fileName":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.registercategory.request.RegisterCategoryRequest":{"type":"object","properties":{"name":{"type":"string"},"parentId":{"type":"integer","format":"int32"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.registercategory.response.CategoryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"parentId":{"type":"integer","format":"int32"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.registercategory.response.ErrorResponse":{"type":"object","properties":{"message":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.registercategory.response.RegisterCategoryResponse":{"type":"object","properties":{"result":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.registercategory.response.ResultResponse"},"error":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.registercategory.response.ErrorResponse"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.registercategory.response.ResultResponse":{"type":"object","properties":{"category":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.registercategory.response.CategoryResponse"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.registerarticle.request.RegisterArticleRequest":{"type":"object","properties":{"title":{"type":"string"},"categoryId":{"type":"integer","format":"int32"},"description":{"type":"string"},"text":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.registerarticle.response.ArticleResponse":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"categoryId":{"type":"integer","format":"int32"},"categoryName":{"type":"string"},"createDateTime":{"type":"string","format":"date-time"},"updateDateTime":{"type":"string","format":"date-time"},"description":{"type":"string"},"text":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.registerarticle.response.ErrorResponse":{"type":"object","properties":{"message":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.registerarticle.response.RegisterArticleResponse":{"type":"object","properties":{"result":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.registerarticle.response.ResultResponse"},"error":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.registerarticle.response.ErrorResponse"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.registerarticle.response.ResultResponse":{"type":"object","properties":{"article":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.registerarticle.response.ArticleResponse"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.getcategory.response.CategoryResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int32"},"name":{"type":"string"},"articleNumber":{"type":"integer","format":"int64"},"parentId":{"type":"integer","format":"int32"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.getcategory.response.ErrorResponse":{"type":"object","properties":{"message":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.getcategory.response.GetCategoryResponse":{"type":"object","properties":{"result":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.getcategory.response.ResultResponse"},"error":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.getcategory.response.ErrorResponse"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.getcategory.response.ResultResponse":{"type":"object","properties":{"categories":{"type":"array","items":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.getcategory.response.CategoryResponse"}}}},"jp.gr.java_conf.simpleblogapi.presentation.article.getarticlebyid.response.ArticleResponse":{"type":"object","properties":{"id":{"type":"string"},"title":{"type":"string"},"categoryId":{"type":"integer","format":"int32"},"categoryName":{"type":"string"},"createDateTime":{"type":"string","format":"date-time"},"updateDateTime":{"type":"string","format":"date-time"},"description":{"type":"string"},"text":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.getarticlebyid.response.ErrorResponse":{"type":"object","properties":{"message":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.getarticlebyid.response.GetArticleByIdResponse":{"type":"object","properties":{"result":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.getarticlebyid.response.ResultResponse"},"error":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.getarticlebyid.response.ErrorResponse"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.getarticlebyid.response.ResultResponse":{"type":"object","properties":{"article":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.getarticlebyid.response.ArticleResponse"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.getarticle.response.ArticleResponse":{"required":["categoryId","categoryName","createDateTime","description","id","text","title","updateDateTime"],"type":"object","properties":{"id":{"type":"string","description":"記事ID"},"title":{"type":"string","description":"記事タイトル"},"categoryId":{"type":"integer","description":"カテゴリID","format":"int32"},"categoryName":{"type":"string","description":"カテゴリ名"},"createDateTime":{"type":"string","description":"作成日時","format":"date-time"},"updateDateTime":{"type":"string","description":"更新日時","format":"date-time"},"description":{"type":"string","description":"記事概要説明"},"text":{"type":"string","description":"記事本文"}},"description":"記事リスト"},"jp.gr.java_conf.simpleblogapi.presentation.article.getarticle.response.ErrorResponse":{"required":["message"],"type":"object","properties":{"message":{"type":"string","description":"エラーメッセージ"}},"description":"エラーレスポンス"},"jp.gr.java_conf.simpleblogapi.presentation.article.getarticle.response.GetArticleResponse":{"required":["error","result"],"type":"object","properties":{"result":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.getarticle.response.ResultResponse"},"error":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.getarticle.response.ErrorResponse"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.getarticle.response.ResultResponse":{"type":"object","properties":{"articles":{"type":"array","description":"記事リスト","items":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.getarticle.response.ArticleResponse"}}},"description":"成功時のレスポンス"},"jp.gr.java_conf.simpleblogapi.presentation.category.deletecategory.response.DeleteCategoryResponse":{"type":"object","properties":{"result":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.deletecategory.response.ResultResponse"},"error":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.category.deletecategory.response.ErrorResponse"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.deletecategory.response.ErrorResponse":{"type":"object","properties":{"message":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.presentation.category.deletecategory.response.ResultResponse":{"type":"object","properties":{"success":{"type":"boolean"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.deletearticle.response.DeleteArticleResponse":{"type":"object","properties":{"result":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.deletearticle.response.ResultResponse"},"error":{"$ref":"#/components/schemas/jp.gr.java_conf.simpleblogapi.presentation.article.deletearticle.response.ErrorResponse"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.deletearticle.response.ErrorResponse":{"type":"object","properties":{"message":{"type":"string"}}},"jp.gr.java_conf.simpleblogapi.presentation.article.deletearticle.response.ResultResponse":{"type":"object","properties":{"success":{"type":"boolean"}}}}}}